Merged PR 173997: Replace git-subtree deploy/cluster with env var VSCLK_CLUSTER_DIR
Managing the git-subtree for deploy/cluster became an impediment while both -core and -cluster are being evolved in tandem. It is easier to keep the repos separate for now.

The interim (or long-term) solution is to introduce the environment variable VSCLK_CLUSTER_DIR where vsclk-core locates its deployment artifacts for the cluster. This pattern can be easily applied to any repo, and doesn't require any specific technology, whether git-subtree, git-submodule, or a package directory.