Merged PR 177074: Remove reading settings from keyvault.
Remove reading settings from keyvault.
The hypothesis here is this is interfering with the rest of app settings as it tries to pull settings from keyvault in the app.
Anthony added this in the App service world, but we never got it to work, removing it as it's not needed here and is likely interfering with our own way of fetching app settings