Merged PR 182952: Support health and warmup controllers in signalr app container
Support new controllers that will allow the cluster to ask for both warmup state and also the general health status of the app.
This PR is the result of a request from @<John Rivard> to provide a mechanism to every service on a kubernetes cluster that can offer 2 entry points to monitor the warmup state and later a health state.
These are not new concepts since Intellicode already use this mechanism and the traffic manager is already performing this actions. More than that these are already part of the vssaas-sdk concepts in terms of interface contracts. Many reference to the saas SDK you will find in this PR.

Attached the email context from John:

Set chart liveness and readiness probe
models/values.yaml
models/templates/service.deployment.yaml

Add health provider in Startup.cs
Models/Startup.cs
…or direct call via SDK:
Vssaas-sdk…HostingServiceCollectionExtensions.AddVsSaaSHosting

Add health and controller endpoint
HealthController.cs
WarmupController.cs

Use async initialization and warmup pattern:
Specify services to be warmed up
Await services to warm up and return OK or InternalServerError

For this particular PR the approach is the following:
- Warmup state: the presence service require 2 running tasks to be completed inn order to consider fully ready to receive a connection into his main hub, these are:
1) Connection to the azure cosmos database for the backplane support
2) Completion of the request to the auth service metadata to be able to correctly validate the tokens being passed form the client. If this 2 tasks are completed the warmup request will unblock and return, otherwise any failed completion will return an intenral status error

- Health state: this task will be performed periodically by the host container, John mention 30 secs as the normal way to invoke this controller, he will later define on the chart what Uri will have this 'health' state. The presence service health controller will monitor 2 things:
1) Successfully completion of the warmup tasks
2) Monitor the state of the 'echo' background task
The design of this PR include a mechanism to test an E2E echo hub so we can fully verify the service is capable of receiving a hub request from a client endpoint.  Since 30 secs seems fast enough we decided to perform an echo task every 5 min but the health controller will return the last known status of echo background task. I already verified this by shutdown the Azure SignalR and so verifying the health will return the internal server status code.

Summary: the PR offer 2 entry new entry points, but the final details on how often they will be called will be tuned up when @<John Rivard> implement the tasks to call this entrypoints.