Merged PR 185662: Improve resiliency for liveness probe end point
With the current logic on the signalr health provider, it check every 5 min the echo hub no matter if the health is true/false.
This change will improve this by reducing the time to 45 secs if the health reports 'unhealthy'.

Also some changes on the liveness probe to have 4 consecutive failures before shutting down the container.

The CLI now has an option to only ping the echo hub to quickly check if an endpoint is in good shape by reporting the  service id and the stamp