Merged PR 186643: Support replication when self contacts register on different regions
There is an important limitation on today's architecture of the signalr service which is that if a contact is registered once in a service it will be not replicated on another region. The assumption is that if you sign on one region, you normally won't register on another service instance or region. We have seen many cases that sometime a machine can have 2 instances of vscode running and each of them on a different region. If that happen no presence consolidation will happen ans also messages will not go trough the backplane.
Several other fixes on telemetry and some adjustment on the throughput for the contacts table.
The contacts table was replicated to avoid collision with the old model