Merged PR 186968: Better support to update the metrics of a presence service
This PR will improve the mechanism to update the metrics of the presence service by providing additional useful properties.
Also the contract was moved into the IBackplaneProvider interface to allow any provider to support the metrics update.

The update background task was moved out form the database backplane provider too.