Merged PR 187544: Ensure OnDisconnectAsync code path always remove the self connection
We have found on our 'prod' deployment that certain contacts when they disconnect they leave a 'stale' self connection and so all of the subscriptions later will wrongly report the status.
We spot a potential problem when the OnDisconnectAsync callback is being called would eventually be cancelled or generate and exception that would prevent the removal of the connection from the self contact.
Additional improvements on the API that retrieve self connections to report the last updated timestamp