Merged PR 187875: Support proper shutdown on the signal service
With recent changes we incorporate a 'unique' lease document collection per service. If there is no proper shutdown the service start/stop will start accumulate unused lease document collection.
This PR will create a way to properly shutdown the signal service and so having the presence service to call dispose on the providers.