Merged PR 188220: Enable AAD authentication on the signalR app service
This PR will allow the sigbalR service to allow receive a AAD token in addition to the existing token validation we get from liveshare.
When an AAD token is validated we need to retrieve the profile from a remote service to normalize the profile id which is how the presence service manage contacts.