Merged PR 189007: Changes to clean 'stale' lease collection created on a previous run
We have made changes in the past to create a unique lease document for each service.
We also add a clean way to delete the collection when the service is shutting down.
But when the pod crash or is interrupted without a clean stop it will leave 'stale' lease collections that needs to be removed and so this PR will add some code during bootstrap to eliminate the document collections that does not have a supported service entry