Merged PR 189500: Support wildcard type notification when adding a subscription to a contact
This PR will allow subscriptions to any property being modified or initialized on a contact.
The previous implementation would force you to specify which properties to monitor and so not suitable to monitor any type of changes that could be passed into another layer.
Modifications on the CLI to exercise this capability and also some unit tests to cover the new options available.