Merged PR 189768: Fix the logger telemetry for the authentication extension
The existing auth extension we have does not goes to our Geneva telemetry due to the creation lifetime.
What happen is that logger factory is being invoked when the telemetry console provider is not yet setup and so the logger will still format the text without the supported json schema.
With this change we fix the logger factory by forcing DI to happen on the 'ConfigureService' and so the logger will be using the telemetry console.
Some other fixes will provide an additional method context to track the perf and use of the auth service Uri.