Merged PR 190303: Add token cache dictionary to avoid multiple request to auth profile service Uri
The signalR service delegate all the authentication tokens into an auth service which will return a profile object.
The problem is that during an successful authentication flow the Jwt handler receive 2 request from the client, one for the 'presencehub/negotiate' and another one for '/presencehub'. There is no need to duplicate the load on the auth service if we can avoid it using a simple cache dictionary for tokens results that are only 30 secs old.