Merged PR 191788: Fix stale self connections on the backplane
This PR will fix the problem detected when many self connections are 'stale' on the backplane cosmos provider due to exceptions that may happen when a connection is disconnected and so our backplane provider is not synched properly.
The solution proposed is to avoid any signalR notifications before synching the backplane and then later proceed with the connection notifications.