Merged PR 194307: stub out new project structure
Stub out new project structure for cloud environments service. Let's use this PR to agree on structure and naming so we can all live happily together under the same `src` dir :wink:

Each root directory is meant to define a POD, one for front end, one for back end, and one for portal. The pods would be co-deployed as a unit. The sub-directories are project/component boundaries that run within that pod.

The legacy folders for `service` and `client` remain for compat. New work goes in new folders so we don't break the old ones.

Let the naming wars begin! :boom:

```
src\
  BackEnd\
    SkuCatalog *TBD
    AzureCapacity
    AzureResourceBroker (Anthony)
    AzureStorage (Derek)
    AzureVirtualMachines (Anu)
    PrivateWebApi (JohnR) *TBD ^ Predominately talks to AzureResourceBroker
  FrontEnd\
    Accounts
    ArmProvider
    Environments (JohnR)
    PublicWebApi (JohnR) ^ Predominately talks to Environments
  Portal\
```