Merged PR 195059: IsHostConnected via Workspace Service.
IsHostConnected via Workspace Service.

When client polls for the status of environment, this service asks workspace to see if the host is connected already.
Sets the status accordingly for the client to handle.

The callback URI still exists and it is still being used. Once IsHostConnected is available on LiveShare prod, the callback can be removed.

Related work items: #944890