Merged PR 195217: Back-end resource ids
I got bored re-watching a movie and decided to code up back-end resource ids. We can debate whether this will be useful throughout the system, and what the final form should take.

The `ResourceId` struct represents the quad { resource type, instance id, subscription id, location  }. The resource type is not to be confused with an azure resource type, but rather, a cloud environment resource type. Perhaps we can think up a better name to avoid downstream confusion.

Both `InstanceId` and `SubscriptionId` are formally of type `Guid` to ensure consistent and valid formats. The `ResourceType` and `Locations` properties are enums to control for the valid set of values and for max string length of the `Id` string.

The `Id` property and the `ToString` method return the canonical format that is useful for passing between the front-end and back-end, and for durable storage. The `Parse` and `TryParse` methods will render the struct from the formatted string.

The Id format is
`vssaas/resourcetypes/{resourcetype}/instances/{instanceid}/subscriptions/{subscriptionid}/locations/{locationid}`

The struct support custom equality, hash code, == and != operators.