Merged PR 196970: Initial commit of start compute
Initial commit of start compute. This follows the continuation pattern we have established. When the `StartCompute` method gets triggered, it places the work on the queue and returns asap. From there it kicks off the start compute process within the compute provider and then as needed, puts tasks on the queue to check for status updates. Independently of that, if `StartCompute` method gets called again with a continuation token, it will do a database lookup to see what the latest status update is.