Merged PR 198104: Add logging to BillingEventManager
- Add logger extensions for logging info related to billing event storage operations
 - Add logging to success and error paths in `BillingEventManager` operations.

In support of that, I did a bit of project refactoring. I split the `Common/HttpContracts` project to separate the logging into a `Common/Common` project, because I didn't want to reference all the HTTP contracts just to be able to do logging. And generally I think we need a place to put utility code that is shared between FE+BE, separate from HTTP contracts. The rest of the changes are cascading effects of that split, mostly updating project references. Also I had to rename the `BackEnd/Common` project to `BackEnd.Common` to avoid collision and confusion of the project name (and to parallel the `FrontEnd.Common` project).