Merged PR 197463: Extract AzureResourceInfo from ResourceId
Replace `ResourceId` with `Guid` as the key for front-end and ResourceBroker to identify back-end cloud environment resources.

Extract `AureResourceInfo` as the means of identifying Azure resources in the compute and storage providers. It contains subscription, resource group, and resource name.

The resource broker's `ResourceRecord`, stored in cosmosdb, now contains the `AzureResourceInfo`. the original `ResourceId` has been removed, as only `base.Id` is required.

With this change, the compute and storage providers no longer use or parse `ResourceId` in any interesting way. These providers return AzureResourceInfo along with their CreateAsync result. The resource broker then stores this in the `ResourceRecord` and is able to look hand it back to the compute or storage provider for subsequent operations.