Merged PR 199495: Changes to exchange the AAD token for a Cascade token. NEW SERVICE
Changes to exchange the AAD token for a Cascade token.

I added a mock class for the new service call, but didn't implement the Boolean since I think that in most scenarios we would need to use the real one for end to end local testing. Perhaps we should add an additional parameter to the app settings.

Related work items: #973423