Merged PR 201255: Make portforwarding ingress work and fix the release
We'd added *.portforwarding.... as a domain in the dev.ci. values but during template substitution the quotes were getting lost and so the helm update was failing. The fix is to add a `{{ . | quote }}` in the chart.

However, since `ingress.hosts` applies to all services, an ingress rule would get added for every service which is not what we want for portforwarding. So I moved the hosts to a `pfingress.hosts` so that only the portal chart can use that value. For the portal, since we need a rewrite annotation to rewrite all paths to one particular endpoint, we need a new Ingress object since annotations are per Ingress object. So I've created a new ingress object such that https://abcdef.portforwarding.online.dev.core.vsengsaas.visualstudio.com/foo?=12102938109238 goes to the configuration page in the portal.

Once we have the actual portforwarding page we can switch to that.