Merged PR 201110: Support Azure redis as a backplane provider
The existing backplane support being used on the presence service is based on Azure Cosmos documents and it has been found that is slow to notify changes happen on different running services, also we could face throttle errors in some cases.
This PR will add a new backplane provider that is based on Azure Redis cache database.
Some refactoring was necessary to define an abstract 'database document' provider and so reuse a lot of code.
Also we are capable to run multiple providers at the same time, so if one fail the other would provide what we need.