Merged PR 201940: VM Assign: Fix for not finding the container image id
VM Assign: Fix for not finding the container image id

We changed the container registry where we store our images which in turn changed the [REPOSITORY[:TAG]] of the image.

Previously, we had specifically looked for "vsclkapps.azurecr.io/kitchensink".

Now, as we know there's only one image in there, we don't specify a filter anymore.

Another approach would be to use "*.azurecr.io/kitchensink" of "*/kitchensink" as the filter but this wouldn't be resilient to changes to the 'kitchensink' part of the image name.