Merged PR 203267: Save location with account info during billing events for billing worker lookup
The location property on account is not being saved during billing event writes.
This is needed for the billing worker to do event lookups by location (data plane location).
I would like to get this checked in so that I can have complete account information to query in dev for testing the billing worker.