Merged PR 204371: [auth]: make the auth service more robust when getting user info
When an external account is used and we aquire the access token to call the ms graph with the `user.read` scope, an `id_token` is returned instead of the `accessToken`. Since `id_token` is not a `JWT` token, we fail when we try to parse it. In that case, since we need only expiration time, we fall back on using the expiration time of the id token instead.