Merged PR 204687: Scaffolding for Windows VM init script
This PR implements the initial scaffolding for the Windows init script. **More work is required on the script itself**.

The custom script for Linux works differently than the custom script for Windows. With the custom script for Linux, you can supply a b64 encoded script to run directly. With custom script for Windows, it's a bit more complicated - you need to store the script you want in Azure storage and point to that location. I've updated the template and how the template args in the PR to reflect the changes need to get this to work.

This raises some complications in how the Windows init script gets stored/updated/versioned. For now, I've created a blob in a new container in the dev storage container and we're just going to point there to get things off the floor. In the future, either some process will need to update this blob, or the script will need to be generic and we pass the script content as an encoded parameter.

The script itself isn't done yet and I'm still working on that. Right now I'm not passing in any of the heartbeat parameters, just the agent blob url and the queue token. The queue token I'm writing to disk although I understand this may need to be an environment variable. I need to test how setting env vars from the custom script works for Windows or if we will need to go with a file based approach. The other part that isn't done is actually starting the VM agent. @<Rajkumar Janakiraman> and I were discussing today we are going to have issues starting VS from the Windows Service, so we may need to run the agent in console mode for now. Regardless, I wanted to get these changes in so we have a script that is running in the dev environment.

Still need to add tests.

Related work items: #984987