Merged PR 205439: [VSO Web] Unblock webviews by using vSCode's fallback domain for webviews
VSCode renders webviews from an isolated domain and we are supposed to pass in the endpoint to use for that. We were passing in localhost which is wrong. In this change, I'm just deleting that so that we fallback to VSCode's default domain for webviews.

Medium term we probably should host those assets ourselves to avoid an extra dependency and also because they are hosted on AWS. I've created #992267 to track that.