Merged PR 204246: Support reading SignalR endpoints from the key vault
Our existing mechanism is not very flexible to define signalR endpoints assigned to a specific region, we basically need to define cluster values for each of the regions. If later we need to add a new connection string for a region that is out of capacity we would need to create a PR and define new values on the chart template and then to a full deployment.
With this mechanism we will create a new SignalR resource for a specific region and then define this new endpoint on our key vault secrets. We still need to restart the service to reconfigure again, but that limitation will go away as soon as the signalR team define a new API to refresh the connections on runtime.
This PR also fix the health hub test and ignore the HTTP 429 error since indicate the signalR connections are out of capacity and so there is no point in restarting the service as we do today. We still generate a log error but with a different method scope so we can easy monitor our telemetry and react quickly on those scenarios.