Merged PR 205564: [VSO Web] Implement a workspace provider to open folders and workspaces
With a recent change in VSCode, OpenFolder\Open Workspace and New Window commands are all broken - they do nothing. We need to implement a `IWorkspaceProvider` and in our case we parse query params for folder , workspace and ew=true (for new window).

Tested that these three commands now work with this provider.