Merged PR 206417: [VSO Web] Implement resolveExternalUri and make portforwarding commands work.
Implements `resolveExternalUri` which will return the full uri for a forwarded port and if the port hasn't been forwaded yet it'll do so. The scenario for the latter is if one clicks on http://localhost:3000 link in the terminal then vscode now calls resolveExternalUri which is supposed to forward the port as well so that opening in the browser actually points to that port.

Also implemented calls to setAuthCookie in reponse to browser sync messages.