Merged PR 206373: Add Traffic Manager host back to portal ingress to fix health probes
Add Traffic Manager host back to portal ingress to fix health probes

I had previously removed the Traffic Manager probes from the Ingress config for the portal service since users are not directly calling from those hosts and the SSL cert is not valid for those hosts, but it turns out that the TM health probes set the TM DNS name in the Host header on health probe requests. This causes the health probe to not match any of the existing host headers that are listed in the Ingress config and ends up returning 404 via the default backend for the health probes.

This change causes the routing to succeed for health probes since the portal service serves up the main landing page for any endpoint (including /healthz which is what the probe is sending).

In the future, we should switch the health probes to a "real" endpoint: either "/" or some static asset that will always exist. "/healthz" currently works, but that it not guaranteed to always be a valid route on the portal/shell.

Related work items: #994157