Merged PR 207558: Scalability fixes
This PR will fix several config values on our service deployment files to properly address the existing issues found on 'euw' stamp, mostly related to memory pressure errors found on  our pods. we increase the replica count to 3 instead of 1 and also we request much more memory to allow the cluster to allocate our main pods in different nodes.
Also we fix the CLI to provide a mechanism to create batch connections and also some enhanced on the telemetry events,