Merged PR 207664: [VSO Web] Clear dotfile localstorage config when a setting's value is empty
This PR simply updates the dotfiles localstorage logic, so that it removes settings when the user clears them as part of an environment creation. Without this change, if I create an environment with a specific dotfile configuration (e.g. setting a repo and install command), than I can never clear it, and I have to manually clear the form fields each time I create a new environment.

I hit this because originally, my dotfiles repo required setting the target path. But I refactored my install script to no longer need it, and I wanted to clear the target path setting.