Merged PR 207701: [Management page] Tactical UX tweaks to improve the "no environments" experience
We'll be demoing VSO a bunch this week, and I just wanted to make a few minor UX tweaks to the management page. These are pretty tactical changes, and when we do the final UX implementation, they'll likely be overriden. That said, these are pretty minor changes, and so can't hurt to make for now :D

This is what the page looks like after these changes:

![image.png](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/207701/attachments/image.png)

This is what it looks like before my changes:

![image.png](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/207701/attachments/image%20(2).png)