Merged PR 207705: [VSO Management page] Minor UX tweaks to the navigation panel
This is one more tactical UX tweak to the navigation panel of the VSO management page. It simply streamlines the spacing a bit, so that it takes up less real-estate on the page, and the link icons are more aligned.

This is what it looks like after my change:

![image.png](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/207705/attachments/image.png)

This is what it looks like before my change:

![image.png](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/207705/attachments/image%20(2).png)