Merged PR 208105: Reintroduce zero cost SKU for static environments
This PR is almost the same as [this one](https://devdiv.visualstudio.com/OnlineServices/_git/vsclk-core/pullrequest/206300?_a=overview), which was rolled back. The main difference is it includes a fix in `InitializeScaleLevelCache.cs`.

**What was the issue in the original PR**
The backend service was attempting to scale up our magic SKU for cloud environments (which does not correspond to an actual VM SKU).

```
    --- Stack Trace ---
    System.InvalidOperationException: Resource Scaling Levels have not been initialized during startup.
     at Microsoft.VsSaaS.Services.CloudEnvironments.ResourceBroker.ResourcePoolDefinitionStore.RetrieveDefinitions() in S:\src\BackEnd\ResourceBroker\Src\ResourceBroker\ResourcePoolDefinitionStore.cs:line 36
     at Microsoft.VsSaaS.Services.CloudEnvironments.ResourceBroker.ResourceBroker.MapLogicalSkuToResourceSku(String skuName, ResourceType type, AzureLocation location)
```

**What was the fix?**
The backend service doesn't need to know about a static environments SKU because billing is in the front end and there is no static environment hardware that we manage. We simply filter out the static environment.

**How was the fix tested?**
I put a breakpoint at the previously thrown exception and ran dev stamps locally. The exception was never thrown, and resources were created successfully. I was also able to create a cloud environment successfully.