Merged PR 208221: Resource Broker - Out of capacity improvements
Ensure that records don't get orphaned when there is no capacity. Further more, support retrying with the existing record when the pool is out of capacity rather than constantly deleting and recreating records. Currently it will retry once per minute until the the `WatchFailedResourcesTask` deleted the record for being not completed within an hour. At that point the record will be removed and another will be added in its place.