Merged PR 208697: Fix the memory leak problem on the Signal Service deployment
This PR will fix the memory problem issue found on the deployed clusters that did restart in around 4 days of operation.
The root of the problem was fixed on another PR on the vsaas-sdk library by forcing the log to be blocking instead of queuing using a batch logger provider. This PR will consume the new nuget package that has the fix.
Additional fixes include more telemetry around memory consumption and also some future preparation of the .NET Core 3 support.

The CLI was enhanced to add an option to send messages to the service