Merged PR 210332: [Bug Fix] Correctly set SKU from deep links when plan isn't cached
In deep links, the `create-environments-panel` can get rendered before the `getPlans` action gets invoked which causes the SKU picker to show "no available instance types" message while the plans load (once plans are loaded it shows them correctly).

The fix here is to also check the `isMadeInitialPlansRequest` flag in addition to `isLoadingPlans` to detect that the panel is rendering before `getPlans` is called and if so, just show the loader instead of the error message.