Merged PR 210327: [Management Portal] Misc. UX tweaks
This PR makes two simple tweaks to the environment creation form:

1. Removing the "required" flag from the `Instance Type` field, since it isn't necessary (drop-down lists with sensible defaults don't require user-interaction)
2. Swapping the SKU and suspend fields, since the suspend option is the more advanced of the two, and I want the form to feature the most commonly used fields

<img src="https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/210327/attachments/image.png" width="300px" />

Additionally, I updated the environment card to account for the fact that the service will currently return a "Provisioning" state for new environments. On the front-end side, we want to always display that as "Creating". Before this change, you would create an environment, and then see it's state as "Creating". However, after a few seconds, the environment's state would update to "Provisioning", which is a little confusing.

<img src="https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/210327/attachments/image%20(2).png" width="300px" />