Merged PR 211131: feat: Added displaying error when creating plans.
feat: Added displaying error when creating plans.

I had to go and get rid of automatic error handling on actions - it relied on us error handling on reducer level but that's not what we want here.

This way - going through the actions we'll have telemetry and will be able to measure failure rates, latency and other goodies.

Related work items: #1009387