Merged PR 211349: [Web editor] Add favicon
This PR simply updates the favicon to use the VS Code red logo (shipped with the VS Code web drop) when the user is in the workbench, and then updates the favicon back to the standard favicon (the MS logo) if they navigate away from it.

<img src="https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/211349/attachments/image.png" width="400px" />

It's possible that I don't need the "unmount" logic, if the user can only get back to the management page via a browser navigation. However, I put it in there for good measure.