Merged PR 211422: [vsow][arm]: fix arm token retrieval for internal users in Edge on windows
Edge on windows performs some dark magic inside the AAD iframe for internal users -- it overrides the AAD redirect with the Javascript redirect that provides different AAD credentials(seem to be the personalized account credentials).  Since we've sandboxed the iframe(and forbidden the JS runtime), their tricks failed. This PR softens the sandboxing a bit and allows script execution inside the iframe hence fixes the issue.

Related work items: #1010832