Merged PR 212043: [Management Portal] Adding SKU specs to their display name
This PR simply changes the display name for SKUs (in the environment card and creation form) to include the CPU/RAM specs as well. That way, devs can quickly identify the meaning of the SKU, without having to jump to the pricing page. For new users, I believe this will also increase the perceived value of the product, since our available options are quite powerful as compared to other services.

<img src="https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/212043/attachments/image%20(2).png" width="300px" />

<img src="https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/212043/attachments/image.png" width="300px" />