Merged PR 211885: add v2 image script to support sig per region
Adding a v2 image script to support having a Shared Image Gallery (SIG) per region. We expect this will allow us to deploy images more quickly. (Creating a separate script so we can continue to work on improving the v1 pipeline while we're bringing up v2. Once v2 is fully onboarded I'll delete the old script and rename this one to the old name.)

If you compare this with the [original script](https://dev.azure.com/devdiv/OnlineServices/_git/vsclk-core?path=%2F.pipelines%2FCopyWindowsImage.ps1&_a=content) you may notice I also removed the use of a temp RG for the image copy. Instead I'm creating the temp Storage Account in the RG containing the SIG.

I also don't decorate the temp SA and container names with target sub name and location because those are already handled by virtue of them being in the target sub and in an RG that is specific to the location.

Related work items: #989763