Merged PR 212437: Increase default plan limit to 1M
Increase default plan limit to 1M

This causes the default plan limit to be increased to 1M if a DB override is not present. Also, on service restart, before the DB value has been loaded, the default config value is used. Since we want to not throttle plan creations anymore, it's good to update the default value in code to prevent some plans from being blocked at service startup prior to the DB cached value overriding the service config.