Merged PR 211716: Independent backplane service support
In our current deployment we define one replica for the presence service support per region to minimize the need to use the global backplane to provide presence state and send messages. The typical time to have a notification flowing into another region is about ~3secs, but with the Redis provider we drastically reduce that time.
In preparation to deploy to VS we except much more connections and traffic and so our need to define more replicas on a cluster deployment is mandatory, but we still want a better solution to traverse different pods when 2 contacts are on different pods but on the same region.
The proposed solution is to define a new .NET Core service that will handle the backplane replication for the whole region and leave the existing service to handle signalR methods and notifications.  To achieve this purpose several refactoring is needed by extracting the backplane logic into a separate assembly that can be used by both services. The new backplane service will handle all the request to replicate a contact on the global scope and also maintain a 'cached' contacts storage to quickly resolve contacts requests and route messages.
A new 'backplane service provider' is implemented to run on each existing service to allow the service to cross other services, both on the same region and global (cross region).

New chart files are defined and incorporated on our build/deploy pipelines.