Merged PR 213561: Put back Azure Cosmos database provider
During the weekend we have continue to see Redis timeout exceptions that could cause the  backplane to provide wrong data for some contacts.
In our current deployment we have just the Redis provider configured and the Cosmos database disabled, it will be better to have both of them enabled and so it one fails we can still replicate the data.
This PR will also fix the serialization when multiple providers are defined and so will invoke the task at the same time before waiting its completion.