Merged PR 213658: fix problems with unique names
I was getting errors about a storage account name already being in use. I _thought_ that account names only had to be unique within a subscription, but based on that behavior I'm thinking they have to be unique within the Azure tenant.

In addition to fixing the storage account problem, I decided to change the snapshot name to use environment instead of subscription name. (Because it's shorter and also consistent with the storage account name.)

Related work items: #989763