Merged PR 214874: add script to update image version in app settings
Added a script to update image version in app settings.

I originally followed @<Derek Bekoe>'s use of the [_File transform_](https://github.com/microsoft/azure-pipelines-tasks/tree/master/Tasks/FileTransformV1) task in the [cloud-environments-service-storage-publish-release](https://devdiv.visualstudio.com/OnlineServices/_releaseDefinition?definitionId=56&_a=definition-tasks&environmentId=189) pipeline.

The problem with that is that it works on the copy of the appsettings.images.json file in the release artifact for vsclk-core, and then uses [MicroBuild's Custom Insertion Command](https://devdiv.visualstudio.com/DevDiv/_wiki/wikis/DevDiv.wiki/632/Execute-Custom-Insertion-Commands) to overwrite the latest version of that file from master and submit that as a PR. Since the release artifact is snapped at about 3.5 hours prior to this update, it's possible for that storage publish pipeline to commit changes to that file and then for the image pipeline to overwrite those changes with the contents it had prior to the storage publish commit.

I had an example of this happening with !214500, which I abandoned before it could do that. In this case, it happened because I was using a topic branch for my vsclk-core artifact. But it could happen even when using master for my artifact if timing of the pipelines is right.

So that's why I'm switching to this script to update the setting.

There's an example of this script getting used [here](https://dev.azure.com/devdiv/OnlineServices/_releaseProgress?_a=release-environment-logs&releaseId=4079&environmentId=21584).

Related work items: #989763